+++ Data/Models/ApplicationUser.cs
@@ -8,5 +8,6 @@ namespace mPole.Data.Models
         public string LastName { get; set; } = string.Empty;
         public DateTime DateOfBirth { get; set; }
         public byte[] ProfileImage { get; set; } = Array.Empty<byte>();
+        public virtual ICollection<Class> Classes { get; set; } = new List<Class>();
     }
 }
