+++ Program.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using mPole.Components;
 using mPole.Components.Account;
-using mPole.Data;
+using mPole.Data.DbContext;
 using mPole.Data.Models;
 using mPole.Data.Repositories;
 using mPole.Interface.Repositories;
