+++ Repository/ImageRepository.cs
@@ -1,4 +1,5 @@
 using Microsoft.EntityFrameworkCore;
+using mPole.Data.DbContext;
 using mPole.Data.Models;
 using mPole.Interface.Repositories;
